requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Supabase
supabase>=2.3.0
postgrest>=0.13.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Payment
razorpay==1.4.2
setuptools>=65.0.0

# Validation
pydantic>=2.9.0
pydantic-settings>=2.5.0
email-validator>=2.2.0

# Utilities
httpx==0.27.0
websockets>=15.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

